CLi Commands added and To Do List Created
Added cd and find to the list. Still need to clarifiy the uses / limits of either command as well as append a more concrete description for ls and cat.

Added a todo list of commands to look into. Those being Grep (which i believe is more akin to a plugin) and any other commands i come across in the future.